Dockerize the project
This adds a Docker env that can be used for running locally in a similar fashion to how it will be run in production. It will create a rails, solr, mysql, and nginx container that communicate with each other. In production, we will be using a persistant instance of solr and mysql, but the architecture will be the same.

Changes to the application to make this work:
- Changed application.rb initialize_on_precompile to stop it from trying to connect to mysql just to precompile assets.
- Pointed the Gemfile at github replicas of hesburgh_assets and hesburgh_infrastructure
- Yet another logger 1.2.8 dependency https://github.com/nahi/logger/issues/3