ESU-1232: Add reindex task
This adds a rake task that can be used to reindex all content. This will
be used on startup within a docker environment to rebuild the index in the
solr container.

Additionally added scripts to handle the nightly import and data cleanup
tasks that were previously handled by the cron jobs defined in
https://github.com/ndlib/ejournal_locator/blob/master/config/schedule.rb.

Replicated the stdout logging behavior in newer rails versions to better
support running in a docker environment (see https://github.com/rails/rails/pull/23734)