Fix robots
Production service is currently not properly serving out the robots.txt,
causing increased traffic to the site. This changes the nginx container
to properly serve out all static assets starting at the /public directory.
The previous config only served out /public/assets. It also changes the
mounting of the directory to use the volumes straight from the rails
container. The previous method involved a little too much magic with
mounting to an external fs and it was causing complications with changing
it to public.