Add rate limits
This adds rate limits to the upstream rails service. Requests to static
assets served by nginx will still be unlimited. This is to protect the
rails service and the services it depends on from being overwhelmed while
ECS is scaling up. Without this, sudden spikes in traffic can cause the
service to fail healthchecks, creating thrashing of tasks. Also adding
support for burst queue to still allow it to spike above the rate that
the app can support in a way that doesn't kill it.